Show actual dice rolls for consumables
This commit removes the `rollConsumable` method from the `DiceRoller`
class. The consumable roll is then replaced with a straight up call to
the normal `roll` method instead. This means the automatic "Succeed" or
"Failed" message is no longer shown and the players need to interpret
the die result themselves. The reasoning for this change is increased
transparency of what was actually rolled. (In my experience the simple
result of success or failure was often treated with suspicion from my
players.)

This commit also adds a dice icon to the consumable bar as an attempted
clue to the fact that the labels are infact clickable to make the
associated roll -- something that is otherwise far from obvious.