Add add features toggles in the weapon sheet
Change the weapon features from a single string to an object containing
all features available in the core rules.
 * melee weapon features: edged, pointed, blunt, parrying and hook
 * ranged weapon feature: slow reload

A free-form list is left to allow custom features to be added to
weapons.

The weapon sheet was modified to let users select the weapon features
via toggles. Available toggles depend on the current weapon's category
(melee or ranged).

There are also some changes in the way weapons are posted to the chat,
mostly to allow the localization of stock weapon features.